requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# HTTP Client
httpx

# Database
neo4j~=5.28.0

# Environment Variables
python-dotenv

# WebSockets
websockets

# Data Validation
pydantic